more lint fixes python sources directory gitignore
